+++ src/core/Plugin.cpp
@@ -129,6 +129,9 @@ namespace OSFUI::Plugin
 			switch (a_msg->type) {
 				case SFSE::MessagingInterface::kPostLoad:
 					REX::INFO("Plugin: SFSE message kPostLoad");
+					if (Runtime::Get().GetConfig().enabled) {
+						Runtime::Get().InstallOverlayDrawPath();
+					}
 					break;
 				case SFSE::MessagingInterface::kPostPostLoad:
 					REX::INFO("Plugin: SFSE message kPostPostLoad");